graph: fix extra bt_put on hashtable removal
Refcounting of the keys is not performed at insertion, we should not do
it at removal.

The keys used in this hashtable are the addresses of the clock_class of
the clock_value. Since each clock_value already holds a reference to the
clock class, there is no need to refcount it.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>