lib: rename "notification" -> "message"
The term "notification" is usually used, in APIs, for asynchronous
alerts of some sort. You get notified for something that you were not
actively waiting for, like an interrupt.

From Merriam-Webster, a _message_ is:

> a communication in writing, in speech, or by signals

and a _communication_ is:

> information communicated : information transmitted or conveyed

Therefore, "message" seems more adapted than "notification" for
Babeltrace's API, where you actively iterate messages which are
communications from an upstream component to a downstream component.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>