Plugin development API: use self enumeration and plugin types
This is similar to a component class method accepting a self component
object.

The self plugin API is empty for the moment, but we might want to add
functions which do not apply to plugin or const plugin objects later
(for example, a function to set a self plugin's private data).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>