Fix: msg-iter.c: accept no packet total and content sizes
This changes an assertion in set_current_packet_content_sizes() to check
that either both the packet and content sizes are set (one can be set
from the other) or both are unset.

It's okay that both total and content sizes are unset: the data decoder
is already implemented to end the packet when the medium returns
`BT_MSG_ITER_MEDIUM_STATUS_EOF`.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>