python-plugin-provider: Make it build
This patch fixes building with --enable-python-plugins.

One problem is that the python plugin provider uses some
functions coming from the internal headers of the lib, which use some
logging functions that are internal to the lib, but not included by
those internal headers.  Trying to add the relevant include file
(lib-logging-internal.h) conflicts with python-plugin-provider's own
logger.

python-plugin-provider doesn't really need to have its own logger.  Even
though it is built as a separate shared object (so that we can package
it separately), it can be considered as internal to the library (and
that's why it includes some internal lib headers).  Therefore, the
easiest way to fix this is to get rid of its logger and make the lib
header files that use lib logging include the required header file
(lib-logging-internal.h).

This requires the python-plugin-provider shared object to access a few
additional symbols not currently exposed by libbabeltrace, which
requires the removal of a few BT_HIDDEN.

I have also included fixes to the python-plugin-provider which I have
on hand.  However, it won't work until the Python bindings are up and
running, so I have marked it as skipped.

Change-Id: I9b46d91fc635d670fef2e4d3c8b32203ec34443f
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/1101
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>