cli: Support arrays in parameters
This patch adds support for arrays in the parameters one can pass to
components on the Babeltrace command line.  For example:

  --params="value=[1,2,\"Hello\"]"

When we are starting to parse a value and the token is a '[', we start
parsing it as an array.  We accept comma-separated values until we reach
the matching ']'. The code should be pretty self-explanatory/commented.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>