Tests: src.ctf.fs: add LTTng packet TS quirks test trace
This trace `even-after-packet` was modified manually to simulate the
hard to reproduce `event-after-packet` LTTng tracers bug.

When the bug occurs, the timestamp of the last event of a packet is larger than
the `timestamp_end` of it's own packet. It makes it look like the event
happened after the packet was closed which is non compliant with the CTF
specification.

To create a trace exhibiting this bug, the `timestamp_end` field of the last
packet was made one clock cycle smaller than the timestamp of the last event of
the said packet.

This bug was fixed in the follow versions of the LTTng tracers:
- lttng-ust 2.11.0
- lttng-module 2.11.0
- lttng-module 2.10.10
- lttng-module 2.9.13

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: Ifd144960fe24ef65a58beb3a8beef4a2def5b0ef