Tests: src.ctf.fs: add lttng-crash packet TS quirk test trace
This trace was modified manually to simulate a trace resulting from the
lttng-crash utility.

When extracting a trace using the lttng-crash utility the `timestamp_end` of
some packets may be left uninitialized with the default value of 0.

To create a trace exhibiting this behavior, the `timestamp_end` of the last
packet was set to 0.

This is not a bug and all LTTng tracers can produce such trace.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: Ic9fa1433bc4cb8b407c7fc50b4a780d2a14d6e69